Add LocalFile abstraction
[Problem]
While goslings and geese may have a local file,
they should also be able to have a representation
that isn't necessarily tied to a local file, as is the
case now.

[Solution]
Remove local file information to a LocalFile struct.
Some information may be copied over to the goose/gosling
objects in time, but it's good to have a separation in
place.