Split functionality across files, attempt ungoose impl
[Problem]
We need to start thinking about splitting up files for
readability. We also must implement the goose function.

[Solution]
Split up functionality by files. Division is not yet
perfect, there's a lot of coupling going on. Attempted
ungoose implementation as well. Part of the reason for
splitting up files is to introduce end-to-end encryption
tests that will make iteration faster.